--- Deartoday/Deartoday/Screen/Main/Controller/MainNavigationController.swift
@@ -1,15 +1,0 @@
-//
-//  MainNavigationController.swift
-//  Deartoday
-//
-//  Created by 이경민 on 2022/07/18.
-//
-
-import UIKit
-
-final class MainNavigationController: UINavigationController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
