--- Deartoday/Deartoday/Screen/Main/Model/MainModel.swift
@@ -1,8 +1,0 @@
-//
-//  MainModel.swift
-//  Deartoday
-//
-//  Created by 소연 on 2022/07/06.
-//
-
-import Foundation
