--- Deartoday/Deartoday/Screen/Main/View/MainView.swift
@@ -1,8 +1,0 @@
-//
-//  MainView.swift
-//  Deartoday
-//
-//  Created by 소연 on 2022/07/06.
-//
-
-import Foundation
